Fixed REST test
